Face mask detector
In this project we have used OpenCV, Keras, and Deep Learning. We introduce a face mask detection model that is based on computer vision. The proposed model can be integrated with surveillance cameras to impede the COVID-19 transmission by allowing the detection of people who are wearing masks not wearing face masks. The model is integration between deep learning and classical machine learning techniques with opencv, tensor flow and keras.